Move requests in FrameworkList to the Frameworks in Project
