+++ src/views/Project/Details/Frameworks/FrameworkList/FrameworkListItem/styles.scss
@@ -1,0 +1,13 @@
+.framework-list-item {
+    display: flex;
+
+    .title {
+        flex-grow: 1;
+    }
+
+    .check {
+        flex-shrink: 0;
+        padding-left: var(--spacing-small);
+        color: var(--color-success);
+    }
+}
