+++ src/views/Project/Details/Frameworks/FrameworkList/index.js
@@ -2,18 +2,10 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Faram from '@togglecorp/faram';
 
-import {
-    RequestCoordinator,
-    RequestClient,
-    requestMethods,
-} from '#request';
-
-import Icon from '#rscg/Icon';
 import SearchInput from '#rsci/SearchInput';
 import SegmentInput from '#rsci/SegmentInput';
 import Checkbox from '#rsci/Checkbox';
 import ListView from '#rscv/List/ListView';
-import ListItem from '#rscv/List/ListItem';
 import AccentButton from '#rsca/Button/AccentButton';
 import LoadingAnimation from '#rscv/LoadingAnimation';
 import modalize from '#rscg/Modalize';
@@ -21,6 +13,7 @@ import modalize from '#rscg/Modalize';
 import _ts from '#ts';
 
 import AddFrameworkModal from './AddFrameworkModal';
+import FrameworkListItem from './FrameworkListItem';
 import styles from './styles.scss';
 
 const AccentModalButton = modalize(AccentButton);
@@ -49,82 +42,14 @@ const fameworkActivityOptions = [
     { key: 'inactive', label: _ts('project.framework', 'frameworkActivityInactiveTitle') },
 ];
 
-// TODO: move to separate component
-const FrameworkListItem = ({
-    className,
-    isActive,
-    isSelected,
-    framework: { title },
-    onClick,
-}) => (
-    <ListItem
-        className={className}
-        active={isActive}
-        onClick={onClick}
-    >
-        <div className={styles.title}>
-            { title }
-        </div>
-        { isSelected &&
-            <Icon
-                name="checkCircle"
-                className={styles.check}
-            />
-        }
-    </ListItem>
-);
-
-FrameworkListItem.propTypes = {
-    className: PropTypes.string,
-    isActive: PropTypes.bool.isRequired,
-    isSelected: PropTypes.bool.isRequired,
-    framework: PropTypes.shape({
-        title: PropTypes.string,
-    }).isRequired,
-    onClick: PropTypes.func.isRequired,
-};
-
-FrameworkListItem.defaultProps = {
-    className: '',
-};
-
 const getFrameworkKey = framework => framework.id;
 
-// TODO: Move request to ../index.js
-const requests = {
-    frameworkListGetRequest: {
-        url: '/analysis-frameworks/',
-        method: requestMethods.GET,
-        query: ({ params: { body } }) => body,
-        onSuccess: ({
-            params: { onSuccess },
-            response,
-        }) => {
-            onSuccess(response);
-        },
-    },
-};
-
-const emptyObject = {};
-const emptyList = [];
-
-@RequestCoordinator
-@RequestClient(requests)
 export default class FrameworkList extends React.PureComponent {
     static propTypes = propTypes;
     static defaultProps = defaultProps;
 
     constructor(props) {
         super(props);
-
-        this.state = {
-            // TODO: move to redux
-            faramValues: {
-                activity: 'active',
-                relatedToMe: true,
-                search: '',
-            },
-        };
 
         this.schema = {
             fields: {
@@ -134,33 +59,6 @@ export default class FrameworkList extends React.PureComponent {
             },
         };
     }
-
-    componentDidMount() {
-        // FIXME: use common function for onFaramChange,
-        const { frameworkListGetRequest } = this.props;
-        const { faramValues } = this.state;
-
-        frameworkListGetRequest.do({
-            body: faramValues,
-            onSuccess: this.handleFrameworkListGetSuccess,
-        });
-    }
-
-    handleFrameworkListGetSuccess = (response) => {
-        console.warn(response);
-    }
-
-    handleFaramChange = (newFaramValues) => {
-        this.setState({ faramValues: newFaramValues });
-
-        const { frameworkListGetRequest } = this.props;
-        const { faramValues } = this.state;
-
-        frameworkListGetRequest.do({
-            body: faramValues,
-            onSuccess: this.handleFrameworkListGetSuccess,
-        });
-    };
 
     itemRendererParams = (key, framework) => ({
         framework,
@@ -176,19 +74,11 @@ export default class FrameworkList extends React.PureComponent {
             projectId,
             setActiveFramework,
             readOnly,
-            frameworkListGetRequest,
+            filterValues,
+            frameworkList,
+            frameworkListPending,
+            onFilterChange,
         } = this.props;
-
-        const { faramValues } = this.state;
-
-        const {
-            pending: frameworkListPending,
-            response: {
-                results: frameworkList = emptyList,
-            } = emptyObject,
-        } = frameworkListGetRequest;
-        // const displayFrameworkList = frameworkList;
-        // console.warn('framework list', frameworkList);
 
         const className = `
             ${classNameFromProps}
@@ -197,8 +87,6 @@ export default class FrameworkList extends React.PureComponent {
 
         return (
             <div className={className}>
-                {/* TODO: Move loading animation to the list view only */}
-                { frameworkListPending && <LoadingAnimation /> }
                 <header className={styles.header}>
                     <div className={styles.top}>
                         <h4 className={styles.heading}>
@@ -222,9 +110,9 @@ export default class FrameworkList extends React.PureComponent {
                     </div>
                     <Faram
                         className={styles.bottom}
-                        onChange={this.handleFaramChange}
+                        onChange={onFilterChange}
                         schema={this.schema}
-                        value={faramValues}
+                        value={filterValues}
                         disable={frameworkListPending}
                     >
                         <SearchInput
@@ -252,6 +140,7 @@ export default class FrameworkList extends React.PureComponent {
                     </Faram>
                 </header>
                 <ListView
+                    pending={frameworkListPending}
                     data={frameworkList}
                     className={styles.content}
                     renderer={FrameworkListItem}
